+++ src/lwipopts.h
@@ -103,4 +103,17 @@
 #define SLIP_DEBUG                  LWIP_DBG_OFF
 #define DHCP_DEBUG                  LWIP_DBG_OFF
 
+// --- SNTP : synchronise l'heure pour la vérification de date des certificats TLS ---
+#define LWIP_DHCP_GET_NTP_SRV       1                       // serveur NTP via DHCP (option 42)
+#define SNTP_SERVER_DNS             1                       // autorise un nom (pool.ntp.org)
+#define SNTP_STARTUP_DELAY          0
+#define SNTP_SET_SYSTEM_TIME(sec)   sntpSetSystemTime(sec)  // callback (voir time_support.h)
+#ifdef __cplusplus
+extern "C" {
+#endif
+void sntpSetSystemTime(unsigned long sec);
+#ifdef __cplusplus
+}
+#endif
+
 #endif /* __LWIPOPTS_H__ */